requirements_test.txt
m3u8~=1.0.0
plexapi~=4.9.1
numpy~=1.22.1
objectrest~=1.1.1
pytest~=6.2.5
python-dotenv~=0.19.2